+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Changes default pitch angle of Search result viewpoint to -35 and add a new uiConfig option to override it: `searchViewpointPitch`.
 - Added a config option to hide panorama footer: `hidePanoramaFooter`
+- Fixes Oblique footer style to match links next to it
 
 ### Plugin Bundle updates